yena first commit

+++ Soomgosusta/src/main/java/kosta/soomgosusta/domain/LinkVO.java
@@ -11,7 +11,7 @@ public class LinkVO {
 	private String l_Status;
 	private String l_Date;
 	private String l_Enter;
-	private int l_Percent;
+	private float l_Percent;
 	private String e_Id;
 	
 	private MemberVO member;
